working on api connecting to modal from greeting button
