+++ components/Home.js
@@ -4,27 +4,27 @@ import GreetingButton from './GreetingButton'
 import { useState, useEffect } from 'react';
 
 const Home = () =>{
-    const [quote, setQuote] = useState('')  
-    const fetchApiCall = () => {
-        return fetch("http://localhost:3001/mantra")
-            .then(response => response.json())
-            .then(data => {
-                setQuote(data)
-            console.log('data: ', data);
-            })
-            .catch(err => {
-            console.log(err);
-            });
-        }
+    // const [quote, setQuote] = useState('')  
+    // const fetchApiCall = () => {
+    //     return fetch("http://localhost:3001/mantra")
+    //         .then(response => response.json())
+    //         .then(data => {
+    //             setQuote(data)
+    //         console.log('data: ', data);
+    //         })
+    //         .catch(err => {
+    //         console.log(err);
+    //         });
+    //     }
 
-        useEffect(() => {
-            fetchApiCall();
-        }, [])
+    //     useEffect(() => {
+    //         fetchApiCall();
+    //     }, [])
 
         return (
             <View style={styles.homeContainer}>
                 <Header />
-                <GreetingButton quote={quote.mantra} fetch={fetchApiCall} title=""/>
+                <GreetingButton />
             </View>
         )
